# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Whispick API — anonymous video recommendation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/w/{token}:
    get:
      operationId: getPublicWhisp
      tags: [public]
      summary: Public recipient landing page data (no auth required)
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public whisp data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicWhisp"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /public/w/{token}/track:
    post:
      operationId: trackWhispEvent
      tags: [public]
      summary: Record a tracking event (opened, clicked, watched)
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingEventInput"
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingResult"
  /public/w/{token}/reply:
    post:
      operationId: publicReply
      tags: [public]
      summary: Recipient sends an anonymous reply from the landing page
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicReplyInput"
      responses:
        "201":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhispReply"
  /public/w/{token}/appreciation:
    post:
      operationId: submitAppreciation
      tags: [public]
      summary: Recipient answers "was this something you needed to hear?"
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppreciationInput"
      responses:
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppreciationResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /public/w/{token}/remind-me:
    post:
      operationId: requestWhispReminder
      tags: [public]
      summary: Recipient schedules a "remind me later" nudge (max 2 total, must land before expiry)
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemindMeInput"
      responses:
        "200":
          description: Reminder scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemindMeResult"
        "400":
          description: Reminder not available (expired, out of reminders, or time out of range)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "410":
          description: Whisp has expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...
    WhispReply:
      type: object
      properties:
        id:
          type: string
        whispId:
          type: string
        replyText:
          type: string
        fromRecipient:
          type: boolean
        videoUrl:
          type: ["string", "null"]
        videoTitle:
          type: ["string", "null"]
        videoThumbnail:
          type: ["string", "null"]
        videoEmbedUrl:
          type: ["string", "null"]
        videoPlatform:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - whispId
        - replyText
        - fromRecipient
        - createdAt
...
    PublicWhisp:
      type: object
      properties:
        id:
          type: string
        videoUrl:
          type: string
        videoTitle:
          type: ["string", "null"]
        videoThumbnail:
          type: ["string", "null"]
        videoEmbedUrl:
          type: ["string", "null"]
        videoStartSeconds:
          type: ["integer", "null"]
        videoPlatform:
          type: ["string", "null"]
        anonymousNote:
          type: ["string", "null"]
        senderAlias:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
        revealRequested:
          type: boolean
        groupSize:
          type: ["integer", "null"]
        appreciationResponse:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        reminderCount:
          type: integer
        expired:
          type: boolean
        hasUpload:
          type: boolean
        aiTakeaway:
          type: ["string", "null"]
        aiTakeawayStatus:
          type: ["string", "null"]
      required:
        - id
        - videoUrl
        - revealRequested
        - expired
        - hasUpload
...
    TrackingEventInput:
      type: object
      properties:
        eventType:
          type: string
      required:
        - eventType
    TrackingResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    PublicReplyInput:
      type: object
      properties:
        replyText:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        videoTitle:
          type: ["string", "null"]
        videoThumbnail:
          type: ["string", "null"]
        videoEmbedUrl:
          type: ["string", "null"]
        videoPlatform:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
...
    AppreciationInput:
      type: object
      properties:
        appreciated:
          type: boolean
      required:
        - appreciated
    AppreciationResult:
      type: object
      properties:
        ok:
          type: boolean
        appreciationResponse:
          type: string
      required:
        - ok
        - appreciationResponse
    RemindMeInput:
      type: object
      properties:
        minutes:
          type: integer
          description: Minutes from now to send the reminder
      required:
        - minutes
    RemindMeResult:
      type: object
      properties:
        ok:
          type: boolean
        nextReminderAt:
          type: string
          format: date-time
        isFinal:
          type: boolean
          description: True if this will be the recipient's last available reminder
      required:
        - ok
        - nextReminderAt
        - isFinal
...