# lib/api-zod/src/generated/types/publicWhisp.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Whispick API — anonymous video recommendation platform
 * OpenAPI spec version: 0.1.0
 */

export interface PublicWhisp {
  id: string;
  videoUrl: string;
  /** @nullable */
  videoTitle?: string | null;
  /** @nullable */
  videoThumbnail?: string | null;
  /** @nullable */
  videoEmbedUrl?: string | null;
  /** @nullable */
  videoStartSeconds?: number | null;
  /** @nullable */
  videoPlatform?: string | null;
  /** @nullable */
  anonymousNote?: string | null;
  /** @nullable */
  senderAlias?: string | null;
  /** @nullable */
  moodTag?: string | null;
  revealRequested: boolean;
  /** @nullable */
  groupSize?: number | null;
  /** @nullable */
  appreciationResponse?: string | null;
  /** @nullable */
  expiresAt?: Date | null;
  reminderCount?: number;
  expired: boolean;
  hasUpload: boolean;
  /** @nullable */
  aiTakeaway?: string | null;
  /** @nullable */
  aiTakeawayStatus?: string | null;
}